# improve stopper performance