## What type of PR is this?

- [ ] API-change
- [ ] BUG
- [x] Improvement
- [ ] Documentation
- [ ] Feature
- [ ] Test and CI
- [ ] Code Refactoring

## Which issue(s) this PR fixes:

issue #14101

## What this PR does / why we need it:

Improves stopper's performance
```
After change:
BenchmarkRunTask1000-8   	     132	   8346828 ns/op
BenchmarkRunTask1000-8   	     132	   8400257 ns/op
BenchmarkRunTask1000-8   	     130	   8416628 ns/op
BenchmarkRunTask1000-8   	     151	   7964632 ns/op
BenchmarkRunTask1000-8   	     138	   7850726 ns/op

Before change:
BenchmarkRunTask1000-8   	      98	  10425915 ns/op
BenchmarkRunTask1000-8   	      80	  12548025 ns/op
BenchmarkRunTask1000-8   	     114	  11788643 ns/op
BenchmarkRunTask1000-8   	     116	   9815367 ns/op
BenchmarkRunTask1000-8   	     100	  10162772 ns/op
```